feat: buyStock 기능 구현

+++ back/src/main/java/dev/pomyharry/stocksimulator/back/model/entity/Account.java
@@ -29,7 +29,7 @@ public class Account {
     private String name;
 
     @Column(name = "DEPOSIT")
-    private int deposit;
+    private long deposit;
 
     @ManyToOne
     @JoinColumn(name = "CUSTOMER_ID")
